adding a cast (double) in a fabs()
git-svn-id: http://ttblue.googlecode.com/svn/branches/active@885 bc2ad3ca-9a4d-0410-8d14-35da9c29b421

+++ includes/core/TTElement.h
@@ -318,7 +318,7 @@ static T TTFold(T value, T low_bound, T high_bound)
 	if((value >= low_bound) && (value <= high_bound))  
 		return value; //nothing to fold 
 	else{
-		foldRange = 2 * fabs(low_bound - high_bound);
+		foldRange = 2 * fabs((double)low_bound - high_bound);
 #ifdef TT_PLATFORM_WIN
 		// The standard remainder() function is not present on Windows, so we do it ourselves.
 		double	v = value - low_bound;
